+++ src/server/js/publish.js
@@ -1,3 +1,0 @@
-/*
- * pwix:accounts/src/server/js/publish.js
- */